requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-experimental>=0.0.39
langchain-ollama>=0.1.4

# Streamlit
streamlit>=1.28.0
streamlit-chat>=0.1.1
streamlit-extras>=0.3.0

# Tools & Utilities
duckduckgo-search>=3.9.4

openai
faiss-cpu
scikit-learn
numpy
httpx
streamlit
pillow
numpy
torch
ultralytics  # Optional: YOLOv9 model runner (if using Ultralytics)
python-dotenv

